eth/tracers: set use callContext to set BlockContext
